move some functions into l10n.js
