+++ packages/wonder-blocks-i18n/src/functions/__tests__/i18n.test.js
@@ -3,15 +3,7 @@ import * as React from "react";
 
 import * as GetLocale from "../get-locale.js";
 import * as FakeTranslate from "../i18n-faketranslate.js";
-import {
-    _,
-    $_,
-    ngettext,
-    doNotTranslate,
-    doNotTranslateYet,
-    localeToFixed,
-    getDecimalSeparator,
-} from "../i18n.js";
+import {_, $_, ngettext, doNotTranslate, doNotTranslateYet} from "../i18n.js";
 
 jest.mock("react", () => {
     return {
@@ -824,166 +816,6 @@ describe("i18n", () => {
 
             // Assert
             expect(result).toEqual("Singular");
-        });
-    });
-
-    describe("# localeToFixed", () => {
-        it("should not call our fake translation", () => {
-            // Arrange
-            // Let's stub out FakeTranslate so we can detect it getting called.
-            const fakeTranslate = {
-                translate: jest.fn().mockReturnValue(""),
-            };
-            jest.spyOn(FakeTranslate, "default").mockImplementation(
-                () => fakeTranslate,
-            );
-
-            // Act
-            localeToFixed(4, 5);
-
-            // Assert
-            expect(fakeTranslate.translate).not.toHaveBeenCalled();
-        });
-
-        describe("default locale (en)", () => {
-            it("should round value up to the given decimal places", () => {
-                // Arrange
-
-                // Act
-                const result = localeToFixed(4.337, 2);
-
-                // Assert
-                expect(result).toEqual("4.34");
-            });
-
-            it("should extend the value to the given decimal places", () => {
-                // Arrange
-
-                // Act
-                const result = localeToFixed(4, 2);
-
-                // Assert
-                expect(result).toEqual("4.00");
-            });
-
-            it("should round the value down to the given decimal places", () => {
-                // Arrange
-
-                // Act
-                const result = localeToFixed(0.0004, 2);
-
-                // Assert
-                expect(result).toEqual("0.00");
-            });
-
-            it("should round negative values and retain the sign", () => {
-                // Arrange
-
-                // Act
-                const result = localeToFixed(-0.0004, 2);
-
-                // Assert
-                expect(result).toEqual("-0.00");
-            });
-
-            it("should round the value to a whole number when 0 decimal places", () => {
-                // Arrange
-
-                // Act
-                const result = localeToFixed(0.0004, 0);
-
-                // Assert
-                expect(result).toEqual("0");
-            });
-
-            it("should round a negative to a whole number when 0 decimal places", () => {
-                // Arrange
-
-                // Act
-                const result = localeToFixed(-0.0004, 0);
-
-                // Assert
-                expect(result).toEqual("0");
-            });
-        });
-    });
-
-    describe("# getDecimalSeparator", () => {
-        it("should handle 'en'", () => {
-            // Arrange
-            jest.spyOn(GetLocale, "getLocale").mockImplementation(() => "en");
-            jest.spyOn(Intl, "NumberFormat").mockImplementation(() => {
-                return {
-                    format: () => "1.1",
-                };
-            });
-
-            // Act
-            const result = getDecimalSeparator();
-
-            // Assert
-            expect(result).toEqual(".");
-        });
-
-        it("should handle 'pl'", () => {
-            // Arrange
-            jest.spyOn(GetLocale, "getLocale").mockImplementation(() => "pl");
-            jest.spyOn(Intl, "NumberFormat").mockImplementation(() => {
-                return {
-                    format: () => "1,1",
-                };
-            });
-
-            // Act
-            const result = getDecimalSeparator();
-
-            // Assert
-            expect(result).toEqual(",");
-        });
-
-        it("should handle 'ar'", () => {
-            // Arrange
-            jest.spyOn(GetLocale, "getLocale").mockImplementation(() => "ar");
-            jest.spyOn(Intl, "NumberFormat").mockImplementation(() => {
-                return {
-                    format: () => "١٫١",
-                };
-            });
-
-            // Act
-            const result = getDecimalSeparator();
-
-            // Assert
-            expect(result).toEqual("٫");
-        });
-
-        it("should handle 'fa-af'", () => {
-            // Arrange
-            jest.spyOn(GetLocale, "getLocale").mockImplementation(
-                () => "fa-af",
-            );
-            jest.spyOn(Intl, "NumberFormat").mockImplementation(() => {
-                return {
-                    format: () => "۱٫۱",
-                };
-            });
-
-            // Act
-            const result = getDecimalSeparator();
-
-            // Assert
-            expect(result).toEqual("٫");
-        });
-
-        it("should handle 'ka'", () => {
-            // Arrange
-            jest.spyOn(GetLocale, "getLocale").mockImplementation(() => "ka");
-
-            // Act
-            const result = getDecimalSeparator();
-
-            // Assert
-            expect(result).toEqual(",");
         });
     });
 });
